Fix linker error in depthwise_conv_op
