+++ tensorflow/core/kernels/depthwise_conv_op.cc
@@ -46,6 +46,7 @@ limitations under the License.
 #endif
 
 #include "tensorflow/core/platform/stream_executor.h"
+#include "tensorflow/core/util/env_var.h"
 #endif  // GOOGLE_CUDA || TENSORFLOW_USE_ROCM
 
 namespace tensorflow {
@@ -264,7 +265,18 @@ extern template struct LaunchDepthwiseConvOp<GPUDevice, Eigen::half>;
 extern template struct LaunchDepthwiseConvOp<GPUDevice, float>;
 extern template struct LaunchDepthwiseConvOp<GPUDevice, double>;
 
-#endif
+bool DisableDepthwiseConvDeterminismExceptions() {
+  static bool cached_disable = [] {
+    bool disable = false;
+    TF_CHECK_OK(tensorflow::ReadBoolFromEnvVar(
+        "TF_DISABLE_DEPTHWISE_CONV_DETERMINISM_EXCEPTIONS",
+        /*default_val*/false, &disable));
+    return disable;
+  }();
+  return cached_disable;
+}
+
+#endif  // GOOGLE_CUDA || TENSORFLOW_USE_ROCM
 
 template <typename Device, typename T>
 class DepthwiseConv2dNativeOp : public BinaryOp<T> {
