+++ tensorflow/core/kernels/depthwise_conv_op.h
@@ -101,7 +101,8 @@ struct LaunchDepthwiseConvBackpropFilterOp<Eigen::GpuDevice, T> {
                   const T* out_backprop, const T* input, T* filter_backprop,
                   TensorFormat data_format);
 };
-#endif
+bool DisableDepthwiseConvDeterminismExceptions();
+#endif  // GOOGLE_CUDA || TENSORFLOW_USE_ROCM
 
 }  // namespace tensorflow
